# website/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RinaDown API",
    "description": "RinaDown 桌面应用的本机 HTTP API。仅监听 `127.0.0.1`（默认端口 17800，可在 设置 → API 服务 中修改）。\n\n- **takeover**（脚本接管）：油猴脚本提交下载，进入快速下载确认流程；需 `X-RinaDown-Client` 头\n- **aria2**：aria2 JSON-RPC 兼容垫片，「发送到 aria2」类脚本与 AriaNg 可直接对接\n- **management**（管理 API）：任务查询/创建/暂停/恢复/删除与队列查询，供 MCP/自动化客户端使用；强制要求 token（`Authorization: Bearer <token>` 或 `X-RinaDown-Token` 头）",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:17800",
      "description": "本机 API 服务（默认端口）"
    }
  ],
  "paths": {
...
    "/api/v1/link/pair/approve": {
      "post": {
        "tags": [
          "link"
        ],
        "summary": "批准/拒绝一次入站配对核验（响应本机收到的 `IncomingPairing` 通知；区别于\n发起方视角、核对 SAS 后调用的 [`api_link_pair_finish`]）。",
        "description": "批准/拒绝一次入站配对核验（响应本机收到的 `IncomingPairing` 通知；区别于发起方视角、核对 SAS 后调用的 pair/finish）。**需 management token**。",
        "operationId": "api_link_pair_approve",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkPairApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "已处理",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkOkResponse"
                }
              }
            }
          },
          "400": {
            "description": "会话不存在/已过期，或载荷非法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultMessage"
                }
              }
            }
          },
          "401": {
            "description": "token 无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "tokenHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LinkOkResponse": {
        "type": "object",
        "description": "通用 `{\"ok\":true}` 应答（发现开关 / 解除配对）。与 [`ResultMessage`] 的\n`{\"success\",\"message\"}` 形态不同——契约就此路由指定的字面 wire 形状。",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
      "LinkPairApproveRequest": {
        "type": "object",
        "description": "批准/拒绝一次入站配对核验（管理面版本；`session_id` 对应\n`LinkEvent{kind:\"incomingPairing\"}` / WS `linkIncomingPairing` 携带的会话 id）。",
        "required": [
          "sessionId",
          "accept"
        ],
        "properties": {
          "accept": {
            "type": "boolean"
          },
          "sessionId": {
            "type": "string"
          }
        }
      },
...
      "ResultMessage": {
        "type": "object",
        "description": "通用结果响应（接管端点应答 / 各端点错误响应统一格式）。",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "tokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RinaDown-Token"
      }
    }
  },
...
}